+++ client/src/io/worxbend/gitea4s/api/PullRequestsApi.scala
@@ -1,8 +1,8 @@
 package io.worxbend.gitea4s.api
 
 import io.worxbend.gitea4s.error.GiteaError
-import io.worxbend.gitea4s.http.PullRequestListParams
-import io.worxbend.gitea4s.model.PullRequest
+import io.worxbend.gitea4s.http.{PullRequestFilesParams, PullRequestListParams}
+import io.worxbend.gitea4s.model.{ChangedFile, PullRequest}
 import zio.{Chunk, IO}
 import zio.stream.ZStream
 
@@ -18,3 +18,10 @@ trait PullRequestsApi:
   def pullRequestByBaseHead(owner: String, repo: String, base: String, head: String): IO[GiteaError, PullRequest]
 
   def pullRequest(owner: String, repo: String, index: Long): IO[GiteaError, PullRequest]
+
+  def pullRequestFiles(
+      owner: String,
+      repo: String,
+      index: Long,
+      params: PullRequestFilesParams = PullRequestFilesParams.default
+  ): ZStream[Any, GiteaError, ChangedFile]